refactor: integrate rules in extractor classes
