+++ src/tab/exportProcessor/scraper/extractor/ExtractorMaxAIGoogle.js
@@ -1,5 +1,5 @@
 import {Extractor} from "./Extractor";
-import {formatLink, setFileHeader} from "../../formatter/formatMarkdown";
+import {formatLink, initTurndown, setFileHeader} from "../../formatter/formatMarkdown";
 
 export default class ExtractorMaxAIGoogle extends Extractor {
   constructor(domain) {
@@ -41,5 +41,51 @@ export default class ExtractorMaxAIGoogle extends Extractor {
 
   getPageSource() {
     return "MaxAI-Google";
+  }
+
+  applyExtractorRules() {
+    initTurndown();
+
+    // turndownConverter.addRule('reformatLinksContainingTags', {
+    //   filter: function (node) {
+    //     return node.nodeName === 'A' && node.getAttribute('href') && node.innerHTML !== node.textContent;
+    //   },
+    //   replacement: function (content, node) {
+    //     let markdown = '';
+    //     for (let i = 0; i < node.childNodes.length; i++) {
+    //       let child = node.childNodes[i];
+    //       if (child.nodeType === Node.ELEMENT_NODE) {
+    //         markdown += turndownConverter.turndown(child.outerHTML); //"<blockquote>" + child.outerHTML + "</blockquote>";
+    //       } else if (child.nodeType === Node.TEXT_NODE) {
+    //         markdown += child.textContent;
+    //       }
+    //     }
+    //     // markdown += "\n" + formatUrl(
+    //     //   node.getAttribute('href').startsWith("/") ?
+    //     //       window.location.protocol + "//" + window.location.host + node.getAttribute('href') :
+    //     //       node.getAttribute('href'),
+    //     //   "Open link ⨠") + "\n";
+    //     return markdown;
+    //   },
+    // });
+    //
+    // turndownConverter.addRule('formatLinks', {
+    //   filter: 'a',
+    //   replacement: function (content, node) {
+    //     const href = node.getAttribute('href');
+    //     const linkText = content.replace(/\\\[/g, '(').replace(/\\\]/g, ')').replace(/</g, '').replace(/>/g, '');
+    //     return '[' + linkText + '](' + href + ')';
+    //   }
+    // });
+    //
+    // turndownConverter.addRule('backslashAngleBracketsNotInBackticks', {
+    //   filter: function (node) {
+    //     return node.querySelectorAll('p').length > 0;
+    //   },
+    //   replacement: function (content, node) {
+    //     // Replace < and > characters in paragraphs but not in backticks
+    //     return "\n" + turndownConverter.turndown(node.innerHTML).replace(/(?<!`)<(?!`)/g, '{{@LT}}').replace(/(?<!`)>(?!`)/g, '{{@GT}}') + "\n\n";
+    //   },
+    // });
   }
 }
