--- src/tab/exportProcessor/scraper/ruler/ruler.js
@@ -1,11 +1,0 @@
-import formatterRules from "./rulesSetter";
-import {dynamicCall} from "../../../../common/utils";
-
-
-/**
- * @description - Set the format rules according to the domain
- * @param domainName
- */
-export function setFormatRules(domainName) {
-  dynamicCall(formatterRules, `set${domainName}Rules`);
-}
