+++ src/tab/exportProcessor/scraper/ruler/rules.js
@@ -1,0 +1,26 @@
+import {turndownConverter} from "../../formatter/formatMarkdown";
+
+export const rules = {
+  'reformatLinksContainingTags': {
+    filter: function (node) {
+      return node.nodeName === 'A' && node.getAttribute('href') && node.innerHTML !== node.textContent;
+    },
+    replacement: function (content, node) {
+      let markdown = '';
+      for (let i = 0; i < node.childNodes.length; i++) {
+        let child = node.childNodes[i];
+        if (child.nodeType === Node.ELEMENT_NODE) {
+          markdown += turndownConverter.turndown(child.outerHTML); //"<blockquote>" + child.outerHTML + "</blockquote>";
+        } else if (child.nodeType === Node.TEXT_NODE) {
+          markdown += child.textContent;
+        }
+      }
+      // markdown += "\n" + formatUrl(
+      //   node.getAttribute('href').startsWith("/") ?
+      //       window.location.protocol + "//" + window.location.host + node.getAttribute('href') :
+      //       node.getAttribute('href'),
+      //   "Open link ⨠") + "\n";
+      return markdown;
+    },
+  }
+}
